pyproject.toml
[tool.poetry]
name = "church_report"
version = "0.1.0"
description = ""
authors = ["kimdooh2019 <choiinhee2020@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "<3.0"
django-extensions = "^2.2.9"
notebook = "^6.0.3"
django-phonenumber-field = "^4.0.0"
phonenumbers = "^8.12.3"
Pillow = "^7.1.2"
djangorestframework = "^3.11.0"
requests = "^2.23.0"
supervisor = "^4.2.0"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
